backend/pyproject.toml
[project]
name = "backend"
version = "4.2.0"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.4.0",
    "dj-database-url>=2.2.0",
    "django>=5.1.1",
    "django-celery-results>=2.5.1",
    "django-prometheus>=2.3.1",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.27.2",
    "gunicorn>=23.0.0",
    "meilisearch>=0.31.5",
    "pillow>=10.4.0",
    "psycopg2-binary>=2.9.9",
    "sentry-sdk>=2.13.0",
    "whitenoise>=6.7.0",
    "django-utils-kit>=1.0.0",
    "django-meilisearch-indexer>=1.0.1",
    "django-anymail[sendinblue]>=13.0",
    "mysql-connector-python",
    "django-soft-delete>=1.0.19",
    "django-cors-headers>=4.4.0",
    "django-filter>=25.1",
    "pandas>=2.3.1",
    "agno>=1.7.7",
    "pydantic>=2.11.7",
    "openai>=1.98.0",
    "httpx>=0.28.1",
    "fastapi>=0.116.1",
    "ag-ui-protocol>=0.1.8",
    "uvicorn>=0.35.0",
]

# ------------------------------
# UV
# ------------------------------
[tool.uv]
dev-dependencies = [
    "coverage>=7.6.1",
    "django-stubs>=5.0.4",
    "djangorestframework-stubs>=3.15.0",
    "factory-boy>=3.3.1",
    "ruff>=0.6.4",
    "tblib>=3.1.0",
    "ty>=0.0.1a3",
    "types-pytz>=2024.1.0.20240417",
    "types-requests>=2.32.0.20240907",
]

# ------------------------------
# TY
# ------------------------------
[tool.ty.rules]
unresolved-attribute = "ignore"

# ------------------------------
# RUFF
# ------------------------------
[tool.ruff.format]
quote-style = "double"
docstring-code-format = true

[tool.ruff.lint.isort]
known-first-party = [
    "core",
    "health",
    "django_react_starter",
    "user",
    "authentication",
]
force-sort-within-sections = true
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]
